+++ RELEASENOTES.md
@@ -1,3 +1,4 @@
 # Admin Module Release Notes
 
+* June 21, 2010 - Fixed eternal redirect bug when an internal access denial tries to redirect (issue #33)
 * June 15, 2010 - **API change** - Refactored functions into separate template and base controllers (issue #20)